feat: extract the version-prefixed postcard codec into a shared tsoracle-codec crate
The paxos and openraft toolkits each carried a near-identical
`[version_byte | postcard(value)]` codec (encode/decode/CodecError). Extract
the framing into a new `tsoracle-codec` micro-crate (mirroring
tsoracle-failpoint / tsoracle-yieldpoint) so it has a single home; both
toolkits' codec modules become thin re-export shims.

SCHEMA_VERSION stays per-toolkit: `version` is a function parameter, so each
toolkit keeps its own constant and evolves its on-disk format independently. A
layout change in one no longer risks a false version mismatch in the other.

The shared CodecError adopts the richer paxos shape (distinct Encode/Decode
variants carrying the postcard error as `#[source]`), a safe widening for the
one external openraft consumer (it matches Version plus a catch-all).

The on-disk byte format is unchanged (golden-byte fixtures pass) and each
toolkit's public codec surface is preserved, so no downstream consumer
changes. paxos-toolkit drops its now-unused direct postcard dependency (it
reaches postcard transitively through tsoracle-codec).

Lands as `feat:` so the new crate publishes before the toolkits that depend on
it.

Closes #265
