+++ Cargo.toml
@@ -15,6 +15,7 @@ members  = [
     "crates/tsoracle-tests",
     "crates/tsoracle-yieldpoint",
     "crates/tsoracle-failpoint",
+    "crates/tsoracle-codec",
     "examples/embedded-server",
     "examples/failover-demo",
     "examples/metrics-prometheus",
@@ -45,6 +46,7 @@ default-members = [
     "crates/tsoracle-paxos-toolkit",
     "crates/tsoracle-yieldpoint",
     "crates/tsoracle-failpoint",
+    "crates/tsoracle-codec",
 ]
 
 [workspace.package]
@@ -90,6 +92,7 @@ tokio              = { version = "1", features = ["macros", "rt-multi-thread", "
 tokio-stream       = { version = "0.1", features = ["sync"] }
 tsoracle-yieldpoint = { path = "crates/tsoracle-yieldpoint", version = "0.1.5" }
 tsoracle-failpoint = { path = "crates/tsoracle-failpoint", version = "0.1.0" }
+tsoracle-codec = { path = "crates/tsoracle-codec", version = "0.1.0" }
 tonic              = "0.14"
 tonic-prost        = "0.14"
 tonic-prost-build  = "0.14"
