+++ crates/tsoracle-codec/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name         = "tsoracle-codec"
+description  = "Shared version-prefixed postcard codec: one `[version_byte | postcard(value)]` framing re-used across the tsoracle toolkits so an on-disk layout change fails loudly instead of misdecoding"
+version      = "0.1.0"
+edition.workspace      = true
+rust-version.workspace = true
+license.workspace      = true
+repository.workspace   = true
+homepage.workspace     = true
+readme.workspace       = true
+authors.workspace      = true
+documentation = "https://docs.rs/tsoracle-codec"
+keywords      = ["serialization", "postcard", "versioning", "codec", "storage"]
+categories    = ["encoding", "database-implementations"]
+
+[dependencies]
+postcard  = { workspace = true }
+serde     = { workspace = true }
+thiserror = { workspace = true }
+
+[dev-dependencies]
+proptest = { workspace = true }
