+++ crates/tsoracle-openraft-toolkit/Cargo.toml
@@ -35,6 +35,7 @@ thiserror        = { workspace = true }
 tokio            = { workspace = true }
 tokio-stream     = { workspace = true }
 tracing          = { workspace = true }
+tsoracle-codec     = { workspace = true }
 tsoracle-failpoint = { workspace = true }
 
 [dev-dependencies]
