+++ crates/tsoracle-paxos-toolkit/Cargo.toml
@@ -29,19 +29,18 @@ futures            = { workspace = true }
 omnipaxos          = { workspace = true, features = ["serde"] }
 parking_lot        = { workspace = true }
 pastey             = { workspace = true }
-postcard           = { workspace = true }
 rocksdb            = { workspace = true, optional = true, features = ["multi-threaded-cf"] }
 serde              = { workspace = true }
 thiserror          = { workspace = true }
 tokio              = { workspace = true }
 tokio-stream       = { workspace = true }
 tracing            = { workspace = true }
+tsoracle-codec     = { workspace = true }
 tsoracle-consensus = { path = "../tsoracle-consensus", version = "0.1.6" }
 tsoracle-core      = { path = "../tsoracle-core", version = "0.2.1" }
 tsoracle-failpoint = { workspace = true }
 
 [dev-dependencies]
-proptest           = { workspace = true }
 rocksdb            = { workspace = true }
 tempfile           = { workspace = true }
 tokio              = { workspace = true, features = ["macros", "rt-multi-thread", "test-util"] }
